# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API",
        "description": "Description of your API",
        "contact": {
            "name": "Contact Name",
            "email": "contact@example.com"
        },
        "license": {
            "name": "License Name",
            "url": "https://www.example.com/license"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products for REST API",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Add a new product to the store",
                "operationId": "addProduct",
                "requestBody": {
                    "description": "Product object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "status",
                                    "types",
                                    "sizes",
                                    "rating"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Example Product Name"
                                    },
                                    "image": {
                                        "description": "Product image file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "pending",
                                            "sold"
                                        ],
                                        "example": "available"
                                    },
                                    "types": {
                                        "type": "string",
                                        "example": "Example Type"
                                    },
                                    "sizes": {
                                        "type": "string",
                                        "example": "Example Size"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product successfully added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product added successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product from the store",
                "description": "This endpoint retrieves a specific product by its ID.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Example Product Name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Example Product Description"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 99.99
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "available",
                                                "pending",
                                                "sold"
                                            ],
                                            "example": "available"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "url/to/product/image"
                                        },
                                        "types": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "type1",
                                                "type2"
                                            ]
                                        },
                                        "sizes": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "small",
                                                "medium",
                                                "large"
                                            ]
                                        },
                                        "rating": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 4.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update an existing product in the store",
                "description": "This endpoint updates an existing product identified by its ID with the provided data.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Product object that needs to be updated in the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Product Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Product Description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "pending",
                                            "sold"
                                        ],
                                        "example": "pending"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "url/to/updated/product/image"
                                    },
                                    "types": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "type1",
                                            "type2"
                                        ]
                                    },
                                    "sizes": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "medium",
                                            "large"
                                        ]
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Updated Product Name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Updated Product Description"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 150
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "url/to/updated/product/image"
                                        },
                                        "types": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "type1",
                                                "type2"
                                            ]
                                        },
                                        "sizes": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "medium",
                                                "large"
                                            ]
                                        },
                                        "rating": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 4.8
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a specific product from the store",
                "description": "This endpoint deletes a specific product by its ID.",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        }
    }
}